Create PriceCalculatorWithDiscount
A simple java example which prompts the user to enter the number of products they intend to buy, prompts the user one by one to enter the price of those items, and if the user purchased more than 5 items they receive a 10% discount. The Price before and after the discount is displayed. If the user does not purchase at least 5 items, then no discount is provided so they will see a message for that and the final price.