Create Arrays.java
A simple program that prompts the user to enter 5 values for an array. It then prints them and their index. Next the program points to the highest value and its index. We then shuffle the values randomly, reprint, and then reverse the order of the random shuffling. Just an example for practice with arrays in java.